+++ NEWS.md
@@ -1,3 +1,11 @@
+V 0.4.1
+=======
+- New features:
+  - New functions:
+    - Functions *target_encode* and *build_target_encoding* have been implemented to provide target encoding which is the process of replacing a categorical value with the aggregation of the target variable.
+  - New features in existing functions : 
+    - Function *prepareSet* integrate *target_encode* function. It is called by providing *target_col* and *target_encoding_functions*.
+
 V 0.4.0
 =======
 - New features: